requirements.txt
python-jose
langchain_groq
fastapi 
uvicorn
pydantic 
python-dotenv
routes
sqlalchemy
PyYAML
langchain
langchain-community
pathlib
streamlit
unstructured
python-docx
langchain_community
sentence-transformers
faiss-cpu